+++ CHANGES.md
@@ -6,6 +6,8 @@ Changes in v1.4.10 (YYYY-MM-DD)
 -------------------------------
 
 - Changed the preferred/first printer URI to use the "ipps" scheme.
+- Updated the USB serial number code to better support non-compliant printers
+  such as those from DYMO (Issue #396)
 - Fixed attribute copying issue from multiple client threads (Issue #390)
 - Now show the default and supported "output-bin" options (Issue #393)
 - Now suppress a duplicate 'auto' value for "media-source-supported" to work